+++ run_gsi_4densvar.sh
@@ -1,3 +1,4 @@
+#!/bin/sh
 echo "Time starting at `date` "
 
 VERBOSE=${VERBOSE:-"YES"}
@@ -263,27 +264,33 @@ if [[ -s $datobs/${prefix_obs}.satwnd.${suffix} ]]; then
 else
    use_prepb_satwnd=.true.
 fi
-SETUP="reduce_diag=.true.,lwrite_peakwt=.true.,lread_obs_save=$lread_obs_save,lread_obs_skip=$lread_obs_skip,l4densvar=.true.,ens_nstarthr=3,iwrtinc=-1,nhr_assimilation=6,nhr_obsbin=$FHOUT,use_prepb_satwnd=$use_prepb_satwnd,lwrite4danl=$lwrite4danl,passive_bc=.true.,newpc4pred=.true.,adp_anglebc=.true.,angord=4,use_edges=.false.,diag_precon=.true.,step_start=1.e-3,emiss_bc=.true.,lobsdiag_forenkf=$lobsdiag_forenkf,lwrite_predterms=.true."
+SETUP="reduce_diag=.true.,lwrite_peakwt=.true.,lread_obs_save=$lread_obs_save,lread_obs_skip=$lread_obs_skip,l4densvar=.true.,ens_nstarthr=3,iwrtinc=-1,nhr_assimilation=6,nhr_obsbin=$FHOUT,use_prepb_satwnd=$use_prepb_satwnd,lwrite4danl=$lwrite4danl,passive_bc=.true.,newpc4pred=.true.,adp_anglebc=.true.,angord=4,use_edges=.false.,diag_precon=.true.,step_start=1.e-3,emiss_bc=.true.,lobsdiag_forenkf=$lobsdiag_forenkf,lwrite_predterms=.true.,thin4d=.true."
 
 if [[ "$HXONLY" = "YES" ]]; then
    #SETUP="$SETUP,lobserver=.true.,l4dvar=.true." # can't use reduce_diag=T
    SETUP="$SETUP,miter=0,niter=1"
 fi
-STRONGOPTS="tlnmc_option=3,nstrong=1,nvmodes_keep=8,period_max=6.,period_width=1.5,baldiag_full=.true.,baldiag_inc=.true.,"
-# no strong bal constraint
-#STRONGOPTS="tlnmc_option=0,nstrong=0,nvmodes_keep=0,baldiag_full=.false.,baldiag_inc=.false.,"
-if [[ "$HXONLY" = "YES" ]]; then
+if [[ "$HXONLY" != "YES" ]]; then
+   if [[ $beta1_inv > 0.999 ]]; then # 3dvar or hybrid gain
+      STRONGOPTS="tlnmc_option=1,nstrong=1,nvmodes_keep=8,period_max=6.,period_width=1.5"
+      SETUP="$SETUP,miter=1,niter(1)=150,niter(2)=0"
+   else # envar
+      STRONGOPTS="tlnmc_option=3,nstrong=1,nvmodes_keep=8,period_max=6.,period_width=1.5,baldiag_full=.true.,baldiag_inc=.true.,"
+      # balance constraint on 3dvar part of envar increment
+      #STRONGOPTS="tlnmc_option=4,nstrong=1,nvmodes_keep=8,period_max=6.,period_width=1.5,baldiag_full=.true.,baldiag_inc=.true.,"
+      # no strong bal constraint
+      #STRONGOPTS="tlnmc_option=0,nstrong=0,nvmodes_keep=0,baldiag_full=.false.,baldiag_inc=.false.,"
+      SETUP="$SETUP,miter=2,niter(1)=50,niter(2)=150"
+   fi
+else
    STRONGOPTS="tlnmc_option=0,nstrong=0,nvmodes_keep=0,baldiag_full=.false.,baldiag_inc=.false.,"
-fi
-if [[ $beta1_inv > 0.999 ]]; then
-   STRONGOPTS="tlnmc_option=1,nstrong=1,nvmodes_keep=8,period_max=6.,period_width=1.5"
 fi
 GRIDOPTS=""
 BKGVERR=""
 ANBKGERR=""
 JCOPTS=""
 #  use tcv_mod, only: init_tcps_errvals,tcp_refps,tcp_width,tcp_ermin,tcp_ermax
-OBSQC="tcp_width=60.0,tcp_ermin=2.0,tcp_ermax=12.0,aircraft_t_bc=$aircraft_bc,biaspredt=1000.0,upd_aircraft=$aircraft_bc" # error variance goes from tcp_ermin (when O-F=0) to tcp_ermax (when O-F=tcp_width=50)
+OBSQC="tcp_width=60.0,tcp_ermin=2.0,tcp_ermax=12.0,aircraft_t_bc=$aircraft_bc,biaspredt=1000.0,upd_aircraft=$aircraft_bc,cleanup_tail=.true." # error variance goes from tcp_ermin (when O-F=0) to tcp_ermax (when O-F=tcp_width=50)
 # GSI defaults
 #   tcp_width=50.0_r_kind
 #   tcp_ermin=0.75_r_kind  
@@ -308,10 +315,10 @@ CHEM=""
 #                   (this is useful for regional application, where there is ambiguity in how to
 #                      define psi,chi from u,v)
 beta1_inv=${beta1_inv:-0.25}
-s_ens_h=${s_ens_h:-800}
-s_ens_v=${s_ens_v:-0.8}
+s_ens_h=${s_ens_h:-400}
+s_ens_v=${s_ens_v:-0.6}
 if [ "$HXONLY" = "NO" ] && [[ $beta1_inv < 0.999 ]]; then
-HYBRIDENSDATA="l_hyb_ens=.true.,n_ens=$nens,beta_s0=$beta1_inv,s_ens_h=$s_ens_h,s_ens_v=$s_ens_v,generate_ens=.false.,uv_hyb_ens=.true.,jcap_ens=$JCAP_ENS,nlat_ens=$NLAT_ENS,nlon_ens=$LONA_ENS,aniso_a_en=.false.,jcap_ens_test=$JCAP_ENS,readin_localization=$readin_localization,write_ens_sprd=.false.,oz_univ_static=.false.,q_hyb_ens=.false.,ens_fast_read=.true."
+HYBRIDENSDATA="l_hyb_ens=.true.,n_ens=$nens,beta_s0=$beta1_inv,s_ens_h=$s_ens_h,s_ens_v=$s_ens_v,generate_ens=.false.,uv_hyb_ens=.true.,jcap_ens=$JCAP_ENS,nlat_ens=$NLAT_ENS,nlon_ens=$LONA_ENS,aniso_a_en=.false.,jcap_ens_test=$JCAP_ENS,readin_localization=$readin_localization,write_ens_sprd=.false.,oz_univ_static=.false.,q_hyb_ens=.false.,ens_fast_read=.true.,readin_beta=$readin_beta"
 else
 HYBRIDENSDATA=""
 fi
@@ -324,16 +331,15 @@ fi
 # Create global_gsi namelist
 cat <<EOF > gsiparm.anl
  &SETUP
-   miter=1,niter(1)=150,niter(2)=0,
    niter_no_qc(1)=25,niter_no_qc(2)=0,
    write_diag(1)=.true.,write_diag(2)=.false.,write_diag(3)=.true.,
    netcdf_diag=.true.,binary_diag=.false.,
    qoption=2,
    factqmin=0.0,factqmax=0.0,deltim=$DELTIM,
-   iguess=-1,
+   tzr_qc=1,iguess=-1,
    oneobtest=.false.,retrieval=.false.,l_foto=.false.,
    use_pbl=.false.,use_compress=.true.,nsig_ext=12,gpstop=50.,
-   use_gfs_nemsio=.true.,sfcnst_comb=.true.,imp_physics=${imp_physics}
+   use_gfs_nemsio=.true.,sfcnst_comb=.true.,cwoption=3,imp_physics=${imp_physics},
    $SETUP
  /
  &GRIDOPTS
@@ -348,6 +354,7 @@ cat <<EOF > gsiparm.anl
    bw=0.0,norsp=4,
    bkgv_flowdep=.false.,bkgv_rewgtfct=1.5,
    bkgv_write=.false.,
+   cwcoveqqcov=.false.,
    $BKGVERR
  /
  &ANBKGERR
@@ -362,8 +369,8 @@ cat <<EOF > gsiparm.anl
    $STRONGOPTS
  /
  &OBSQC
-   dfact=0.75,dfact1=3.0,noiqc=.true.,oberrflg=.true.,c_varqc=0.02,
-   use_poq7=.true.,qc_noirjaco3_pole=.true.,
+   dfact=0.75,dfact1=3.0,noiqc=.true.,oberrflg=.false.,c_varqc=0.02,
+   use_poq7=.true.,qc_noirjaco3_pole=.true.,vqc=.true.,
    $OBSQC
  /
  /
@@ -816,7 +823,7 @@ ls -l
 echo "Time before GSI `date` "
 export PGM=$tmpdir/gsi.x
 export FORT_BUFFERED=TRUE
-sh ${enkfscripts}/runmpi
+${enkfscripts}/runmpi
 rc=$?
 if [[ $rc -ne 0 ]];then
   echo "GSI failed with exit code $rc"
@@ -932,7 +939,7 @@ fi
 #             cat $HOSTFILE
 #          fi
 #          corecount=$((corecount+$nprocs))
-#          sh ${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out 2> nc_diag_cat_${type}_${string}.out &
+#          ${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out 2> nc_diag_cat_${type}_${string}.out &
 #          if [ $corecount -eq $cores ]; then
 #             echo "waiting... corecount=$corecount"
 #             wait
@@ -969,7 +976,7 @@ for loop in $loops; do
             export PGM="${execdir}/nc_diag_cat.x -o ${savdir}/diag_${type}_${string}.${adate}_${charnanal2}.nc4  pe*${type}_${loop}*nc4"
             ls -l pe*${type}_${loop}*nc4
             nodecount=$((nodecount+1))
-            if [ "$machine" == 'theia' ]; then
+            if [ ! -z $SLURM_JOB_ID ] && [ "$machine" == 'theia' ]; then
                node=`head -$nodecount $NODEFILE | tail -1`
                export HOSTFILE=hostfile_${nodecount}
                /bin/rm -f $HOSTFILE
@@ -981,8 +988,9 @@ for loop in $loops; do
                echo "contents of hostfile_${nodecount}..."
                cat $HOSTFILE
             fi
-            sh ${enkfscripts}/runmpi 1> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.out 2> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.err &
-            #sh ${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out &
+            ${enkfscripts}/runmpi 1> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.out 2> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.err &
+            #${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out &
+            sleep 1
             if [ $nodecount -eq $totnodes ]; then
                echo "waiting... nodecount=$nodecount"
                wait
